requirements.txt
streamlit
google.generativeai
pathlib
requests
pandas

